Moved slots module into baselib
