+++ openquake/baselib/slots.py
