+++ openquake/hazardlib/mfd/truncated_gr.py
@@ -22,7 +22,7 @@ import math
 
 from openquake.baselib.python3compat import range, round
 from openquake.hazardlib.mfd.base import BaseMFD
-from openquake.hazardlib.slots import with_slots
+from openquake.baselib.slots import with_slots
 
 
 @with_slots
