+++ openquake/hazardlib/scalerel/peer.py
@@ -17,7 +17,7 @@
 Module :mod:`openquake.hazardlib.scalerel.peer` implements :class:`PeerMSR`.
 """
 from openquake.hazardlib.scalerel.base import BaseMSRSigma
-from openquake.hazardlib.slots import with_slots
+from openquake.baselib.slots import with_slots
 
 
 @with_slots
